requirements.txt
Flask==1.1.1
Flask_Bcrypt==0.7.1
python-dotenv==0.12.0
